requirements.txt
absl-py==0.4.1
astor==0.7.1
audioread==2.1.6
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
future==0.16.0
gast==0.2.0
grpcio==1.15.0
horovod==0.13.10
idna==2.7
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
librosa==0.6.1
llvmlite==0.24.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.4
mock==2.0.0
mpi4py==3.0.0
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.2.5
notebook==5.7.0
numba==0.39.0
numpy==1.14.5
nvidia-dali==0.2.0
pandas==0.23.0
pandocfilters==1.4.2
parso==0.3.1
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.5
plotly==3.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.7
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-speech-features==0.6
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.20.0
resampy==0.2.1
retrying==1.3.3
sacrebleu==1.2.11
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
sentencepiece==0.1.4
six==1.11.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0+nv
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
typing==3.6.6
urllib3==1.24
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
